requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aioresponses==0.7.6
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
backoff==2.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
fastapi==0.115.0
fonttools==4.54.1
frozenlist==1.4.1
gotrue==2.9.2
greenlet==3.1.1
grpcio==1.66.2
grpcio-tools==1.66.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
inference-sdk==0.9.17
jiter==0.6.1
kiwisolver==1.4.7
Mako==1.3.5
MarkupSafe==3.0.1
marshmallow==3.22.0
matplotlib==3.9.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==2.1.2
openai==1.51.2
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
portalocker==2.10.1
postgrest==0.17.1
propcache==0.2.0
protobuf==5.28.2
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
qdrant-client==1.12.0
realtime==2.0.5
requests==2.32.3
scipy==1.14.1
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
storage3==0.8.1
StrEnum==0.4.15
supabase==2.9.0
supafunc==0.6.1
supervision==0.24.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.0
websockets==13.1
yarl==1.14.0
